requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
plotly-express
jupyter==1.0.0
notebook==6.4.4
xgboost==2.1.4
optuna==4.1.0
PyYAML
tqdm
